+++ test/vector_prop_test.exs
@@ -236,6 +236,12 @@ defmodule A.Vector.PropTest do
       assert false ===
                A.Vector.any?(A.Vector.new(filtered_list), fn x -> !hash_multiple_of_2(x) end)
 
+      assert Enum.find(list, &hash_multiple_of_2/1) ===
+               A.Vector.find(vector, &hash_multiple_of_2/1)
+
+      assert Enum.find_value(list, &hash_multiple_of_2/1) ===
+               A.Vector.find_value(vector, &hash_multiple_of_2/1)
+
       assert capture_error(Enum.sum(list)) === capture_error(A.Vector.sum(vector))
 
       assert capture_error(Enum.reduce(list, 1, &(&2 * &1))) ===